pyproject.toml
[tool.poetry]
name = "home-automations"
version = "0.0.0"
description = "My custom home automation software written in Python."
authors = ["Lennard Beers <l.beers@outlook.de>"]
license = "MIT"
readme = "README.md"
packages = [
    { include = "home_automations" },
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
home-automations = "home_automations.__main__:start"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
python-dotenv = "^1.0.0"
pyyaml = "^6.0.1"
marshmallow-dataclass = "^8.6.0"
pytz = "^2023.3.post1"
python-hass-client = {git = "https://github.com/EuleMitKeule/python-hass-client", rev = "master"}
pytibber = "^0.28.2"
colour = "^0.1.5"
types-pytz = "^2023.3.1.1"
types-pyyaml = "^6.0.12.12"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
hypercorn = "^0.16.0"

[tool.poetry.group.dev.dependencies]
black = "^23.12.1"
ruff = "^0.1.11"
mypy = "^1.8.0"
pre-commit = "^3.6.0"
homeassistant = "^2024.1.5"

[tool.poetry.group.test.dependencies]
pytest = "^7.4.4"

[tool.ruff]
unfixable = ["F401"]

[tool.mypy]
ignore_missing_imports = true
check_untyped_defs = true
